Full scan reinitialise.
The scan on/off code worked before, but provided nothing interfered
with the HCI state. The thing is using the HCI for anything other than
scanning (e.g. connecting to a device) interferes with the state. That
made the stop/start code functionally useless.

Oddly enough the symptom was that it could receive scan events but
only from the device just connected to (at least on the RPi). I've not
investigated further yet, but there seem to be filters in several differnt
locations which filter events by address. That might have something to do
with it.